#8 Added Maven build files.
Note that Maven1 is required to build the plugins. There seems to be a
possibility to use Maven2 or newer, but I can't manage to digest the
documentation, as it seems quite more complicated. Also, it is assumed
that the updatesite is one directory upwards in a directory called
"update". So the directory structure should look like this:

    root
      |-ecd
      |   \-pom.xml  <- Invoke Maven here.
      \-update
          \-pom.xml

The main build command (`mvn clean package`) must be invoked from the
"ecd" directory.